🌱 [scanner] refactor: split large drill-down views (Part of #21375)
- ClusterDrillDown.tsx: 1003 → 661 lines (-34%)
  Extract ClusterResourceTree component to cluster-drilldown/ subdir
  Move TreeLens and ClusterTab types to cluster-drilldown/types.ts
  Remove eslint-disable max-lines comment

- DeploymentDrillDown.tsx: 901 → 789 lines (-12%)
  Extract buildLabelSelector, classifyScaleError helpers to deployment-drilldown/helpers.ts
  Extract TabType, Props, MAX_SCALE_REPLICAS, POD_STATUS_CONFIG to deployment-drilldown/types.ts
  Remove eslint-disable max-lines comment
