+++ web/src/components/drilldown/views/cluster-drilldown/index.ts
@@ -1,0 +1,3 @@
+export type { TreeLens, ClusterTab } from './types'
+export { ClusterResourceTree } from './ClusterResourceTree'
+export type { ClusterResourceTreeProps } from './ClusterResourceTree'
