+++ web/src/components/drilldown/views/cluster-drilldown/types.ts
@@ -1,0 +1,5 @@
+/** Lens/view filter options for the cluster resource tree */
+export type TreeLens = 'all' | 'issues' | 'nodes' | 'workloads' | 'storage' | 'network'
+
+/** Tab IDs for the ClusterDrillDown bottom section */
+export type ClusterTab = 'events' | 'resources'
