+++ web/src/components/drilldown/views/deployment-drilldown/index.ts
@@ -1,0 +1,4 @@
+export type { TabType, Props } from './types'
+export { MAX_SCALE_REPLICAS, POD_STATUS_CONFIG } from './types'
+export type { LabelSelectorRequirement, ScaleErrorKey } from './helpers'
+export { classifyScaleError, buildLabelSelector } from './helpers'
